+++ CHANGELOG.rst
@@ -63,6 +63,9 @@ Added
   `#9 <https://github.com/openego/eGon-data/issues/9>`_
 * Add hydro and biomass power plants eGon2035
   `#127 <https://github.com/openego/eGon-data/issues/127>`_
+* Creation of the ehv/hv grid model with osmTGmod, see 
+  `issue #4 <https://github.com/openego/eGon-data/issues/4>`_ and 
+  `PR #164 <https://github.com/openego/eGon-data/pull/164>`_  
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159